pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.95.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
